requirements.txt
qrcode==7.4.2
pillow==10.0.0
pyzbar==0.1.9
opencv-python==4.7.0.72
reportlab==4.0.4
ttkbootstrap==1.10.1
tkcalendar==1.6.1
openpyxl==3.1.2
pyinstaller==5.13.0
fpdf2==2.7.5